Issue #12186 - Fix duplicated function and move create Contributors to Trait
